Code organization
Fixes #1
